+++ tinper-acs/ac-grids/CHANGELOG.md
@@ -1,3 +1,38 @@
+<a name="0.1.7"></a>
+## [0.1.7](https://github.com/tinper-bee/ac-grids/compare/v0.1.6...v0.1.7) (2019-09-23)
+
+
+
+<a name="0.1.6"></a>
+## [0.1.6](https://github.com/tinper-bee/ac-grids/compare/v0.1.5...v0.1.6) (2019-09-16)
+
+
+
+<a name="0.1.5"></a>
+## [0.1.5](https://github.com/tinper-bee/ac-grids/compare/v0.1.4...v0.1.5) (2019-09-16)
+
+
+
+<a name="0.1.4"></a>
+## [0.1.4](https://github.com/tinper-bee/ac-grids/compare/v0.1.3...v0.1.4) (2019-09-11)
+
+
+
+<a name="0.1.3"></a>
+## [0.1.3](https://github.com/tinper-bee/ac-grids/compare/v0.1.2...v0.1.3) (2019-09-09)
+
+
+
+<a name="0.1.2"></a>
+## [0.1.2](https://github.com/tinper-bee/ac-grids/compare/v0.1.1...v0.1.2) (2019-09-06)
+
+
+### Features
+
+* **ac-grids:** 升级bee-table ([caafe0d](https://github.com/tinper-bee/ac-grids/commit/caafe0d))
+
+
+
 <a name="0.1.1"></a>
 ## [0.1.1](https://github.com/tinper-bee/ac-grids/compare/v0.1.0...v0.1.1) (2019-09-04)