+++ tinper-acs/ac-search-cn/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="0.0.28"></a>
+## [0.0.28](https://github.com/tinper-bee/ac-search-cn/compare/v0.0.27...v0.0.28) (2019-09-16)
+
+
+
 <a name="0.0.27"></a>
 ## [0.0.27](https://github.com/tinper-bee/ac-search-cn/compare/v0.0.26...v0.0.27) (2019-08-26)