+++ tinper-acs/ac-split-area/CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="0.0.4"></a>
+## [0.0.4](https://github.com/tinper-bee/ac-split-area/compare/v0.0.3...v0.0.4) (2019-09-23)
+
+
+### Features
+
+* **ac-split-area:** css ([674af3e](https://github.com/tinper-bee/ac-split-area/commit/674af3e))
+
+
+
 <a name="0.0.3"></a>
 ## [0.0.3](https://github.com/tinper-bee/ac-split-area/compare/v0.0.2...v0.0.3) (2019-08-20)