+++ tinper-acs/ac-split-button/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="0.0.6"></a>
+## [0.0.6](https://github.com/tinper-bee/ac-split-button/compare/v0.0.5...v0.0.6) (2019-09-11)
+
+
+
 <a name="0.0.5"></a>
 ## [0.0.5](https://github.com/tinper-bee/ac-split-button/compare/v0.0.4...v0.0.5) (2019-08-29)